+++ src/app/js/admin/admin.js
@@ -1,0 +1,17 @@
+(function() {
+  'use strict';
+
+  angular.module('orrportal.admin', [
+    ,'orrportal.admin.users'
+    ,'orrportal.admin.orgs'
+
+  ])
+    .directive('adminMenu', function() {
+      return {
+        restrict:    'E',
+        templateUrl: 'js/admin/adminmenu.html'
+      }
+    })
+  ;
+
+})();